Adjusted ParkingSpot to be able to contain multiple vehicles. Updated ParkVehicle method and PrintGarage method

+++ Prague Parking 2.0/Vehicle.cs
@@ -2,27 +2,25 @@
 
 public interface ICalculateParkingCost
 {
-    public double CalculateParkingCost(double Space, int Rate, DateTime ParkingStartTime);
+    public double CalculateParkingCost(int Space, int Rate, DateTime ParkingStartTime);
 }
 public abstract class Vehicle : ICalculateParkingCost
 {
     public abstract string? RegNumber { get; set; }
     public abstract string Type { get; set; }
-    public abstract double Rate { get; set; }
-    public abstract double Space { get; set; }
+    public abstract int Rate { get; set; }
+    public abstract int Space { get; set; }
     public DateTime ParkingStartTime { get; set; }
 
     public Vehicle()
     {
 
     }
-    public double CalculateParkingCost(double Space, int Rate, DateTime ParkingStartTime)
+    public double CalculateParkingCost(int Space, int Rate, DateTime ParkingStartTime)
     {
         TimeSpan parkedDuration = DateTime.Now - ParkingStartTime;
         return Space * Rate * parkedDuration.TotalHours;
     }
-
-
 }
 
 
@@ -30,8 +28,8 @@ class Bike : Vehicle
 {
     public override string? RegNumber { get; set; }
     public override string Type { get; set; } = "Bike";
-    public override double Rate { get; set; } = 5;
-    public override double Space { get; set; } = 1;
+    public override int Rate { get; set; } = 5;
+    public override int Space { get; set; } = 1;
 
     public Bike(string regnumber)
     {
@@ -46,8 +44,8 @@ class MC : Vehicle
 {
     public override string? RegNumber { get; set; }
     public override string Type { get; set; } = "MC";
-    public override double Rate { get; set; } = 10;
-    public override double Space { get; set; } = 2;
+    public override int Rate { get; set; } = 10;
+    public override int Space { get; set; } = 2;
 
     public MC(string regnumber)
     {
@@ -60,8 +58,8 @@ class Car : Vehicle
 {
     public override string? RegNumber { get; set; }
     public override string Type { get; set; } = "Car";
-    public override double Rate { get; set; } = 20;
-    public override double Space { get; set; } = 4;
+    public override int Rate { get; set; } = 20;
+    public override int Space { get; set; } = 4;
 
     public Car(string regnumber)
     {
@@ -74,8 +72,8 @@ class Bus : Vehicle
 {
     public override string? RegNumber { get; set; }
     public override string Type { get; set; } = "Bus";
-    public override double Rate { get; set; } = 40;
-    public override double Space { get; set; } = 16;
+    public override int Rate { get; set; } = 40;
+    public override int Space { get; set; } = 16;
 
     public Bus(string regnumber)
     {
